### 2025-08-28 #8884 @danielskatz
👋 @chemsh - thanks for your submission.

I note that it is about 50% longer than JOSS recommends. While some of this is due to code being counted as words, please still consider if you can shorten it. You can use the commands `@editorialbot check repository` to check a few things that includes the word count, and the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, while this is otherwise ready to be assigned to an editor, we don't have a suitable one available for the minute, so I'm going to mark this as waitlisted for now.
