### 2025-09-06 #8884 @HaoZeke
hi @marshallward @awvwgk @DimitriPapadopoulos @pearu @jameskermode @mgoonde @certik 👋 would (any/all of) you be interested in and available to review this JOSS submission? I was thinking it would be of interest due to all the experience working in and around Fortran interop :)
