### 2025-09-06 #8884 @awvwgk
Looks interesting, I am up for reviewing this.
