### 2025-09-14 #8884 @HaoZeke
Going ahead with this now, but others can be added if they feel the need :)
