### 2025-09-14 #8992 @editorialbot
[REVIEW]: DL_PY2F: A library for Python-Fortran interoperability
**Submitting author:** @chemsh (<a href="http://orcid.org/0000-0002-7524-4179">You Lu</a>)
**Repository:** https://github.com/stfc/dl_py2f
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0
