### 2025-10-23 stfc/dl_py2f#19 @mgoonde
[JOSS REVIEW]
Hello,
these are my comments for the review at https://github.com/openjournals/joss-reviews/issues/8992

On several places in the paper it is claimed that the data in `DL_PY2F` is mapped to Fortran pointers, and its retrieval does not create any copies, in particular for the arrays.

My first concern is that this claim is not true.
I have placed some statements to write out the memory location of variables (`loc()`) into `dl_py2f.f90`, and it seems there are copies of data being made within the getter routines (for example `returnTwoDimDbl`). Please clarify this point.

My second concern is the speed of execution of the example provided in https://github.com/stfc/dl_py2f-example.
The example takes a full 12 seconds to finish on my laptop, while it does not seem to do much more than passing around the data. Please correct me if this is wrong.

At the end of the section in the paper:
> Comparison with other tools

it is claimed that:
> " ... and is portable across modern Fortran compilers";

meanwhile the preceding sections talks strictly about GNU compiler, due to the format of `.mod` files being proprietary, and the prior text mentions only Intel as additionally tested compiler. Please clarify your claim.
