### 2025-11-28 4e6da29
Refactored the code to make the Python-to-Fortran fully memory-duplication-free
In the previous version although DL_PY2F itself did not duplicate the original NumPy arrays by using only pointers, the application developer (namely, DL_PY2F user) still needed to allocate arrays which were duplications in memory.

In the latest update, arrays of a pointer attribute have read/write access to the data of NumPy arrays through the %get() function, e.g.:
real(kind=8), pointer:: arary1(:)
PyObj%get('array1', array1)
! this changes the value of obj.array1[0] on the Python side
array1(1) = 1.2345

For sa
[…]
