+++ dl_py2f.f90
@@ -50,6 +50,8 @@ module DL_PY2F
         class(*)        , pointer :: array(:)        => null()
         type(PyType)    , pointer :: PyPtr           => null()
         character(len=8), pointer :: onedimchar(:)   => null()
+        integer(kind=4) , pointer :: onedimint(:)    => null()
+        integer(kind=4) , pointer :: twodimint(:,:)  => null()
         integer(kind=8) , pointer :: onedimlng(:)    => null()
         integer(kind=8) , pointer :: twodimlng(:,:)  => null()
         real(kind=8)    , pointer :: onedimdbl(:)    => null()
@@ -63,17 +65,19 @@ module DL_PY2F
                                      assignOneDimDbl,  &
                                      assignTwoDimDbl,  &
                                      assignOneDimInt,  &
-                                     assignTwoDimInt
+                                     assignTwoDimInt,  &
+                                     assignOneDimLng,  &
+                                     assignTwoDimLng
         generic, public :: get    => getInt,        &
                                      getIntCLong,   &
                                      getReal,       &
                                      getDouble,     &
-                                     getOneDimInt,  &
-                                     getOneDimLong, &
                                      getOneDimDbl,  &
-                                     getTwoDimInt,  &
-                                     getTwoDimLong, &
                                      getTwoDimDbl,  &
+                                     getOneDimInt,  &
+                                     getTwoDimInt,  &
+                                     getOneDimLng,  &
+                                     getTwoDimLng,  &
                                      getChar,       &
                                      getCCharArray, &
                                      getLogical,    &
@@ -84,8 +88,8 @@ module DL_PY2F
                                      setTwoDimDbl,  &
                                      setOneDimInt,  &
                                      setTwoDimInt,  &
-                                     setOneDimLong, &
-                                     setTwoDimLong, &
+                                     setOneDimLng,  &
+                                     setTwoDimLng,  &
                                      setCCharArray, &
                                      setScalar
         procedure, public :: initialise => private_initialise
@@ -93,21 +97,23 @@ module DL_PY2F
         procedure, public :: finalise   => private_finalise
         procedure, private :: assignScalar
         procedure, private :: assignOneDimChar
-        procedure, private :: assignOneDimInt
-        procedure, private :: assignTwoDimInt
         procedure, private :: assignOneDimDbl
         procedure, private :: assignTwoDimDbl
+        procedure, private :: assignOneDimInt
+        procedure, private :: assignTwoDimInt
+        procedure, private :: assignOneDimLng
+        procedure, private :: assignTwoDimLng
         procedure, private :: assignPyPtr
         procedure, private :: getIntCLong
         procedure, private :: getInt
         procedure, private :: getReal
         procedure, private :: getDouble
-        procedure, private :: getOneDimInt
-        procedure, private :: getOneDimLong
         procedure, private :: getOneDimDbl
-        procedure, private :: getTwoDimInt
-        procedure, private :: getTwoDimLong
         procedure, private :: getTwoDimDbl
+        procedure, private :: getOneDimInt
+        procedure, private :: getTwoDimInt
+        procedure, private :: getOneDimLng
+        procedure, private :: getTwoDimLng
         procedure, private :: getChar
         procedure, private :: getCCharArray
         procedure, private :: getLogical
@@ -118,18 +124,20 @@ module DL_PY2F
         procedure, private :: setTwoDimDbl
         procedure, private :: setOneDimInt
         procedure, private :: setTwoDimInt
[…]
