### 2025-12-03 1d77ad1
Refactored dl_f2py.py by greatly improving its performance
- Adapted to a lazy-construction mode which instantiates <class DL_DT> on the fly only upon access requests, instead of the prevous full construction of all Python entities
- Enabled PDE-style slicing of arrays
- Enabled other NumPy-style indexing using such as negative integers and None
- Added cache mechanisms to further optimise the performance (minor)

As a result it runs more than 30 times faster than before.
