+++ dl_f2py.py
@@ -19,15 +19,20 @@
 # you.lu@stfc.ac.uk
 __email__ = 'you.lu@stfc.ac.uk'
 __author__ = f'You Lu <{__email__}>'
-import utils
-import os
+import libpath, os, utils
 from utils.colours import *
-from ctypes  import addressof, CDLL, memset, RTLD_GLOBAL, sizeof
-from ctypes  import c_bool, c_char_p, c_double, c_float, c_int, c_long, c_void_p, POINTER
-from ctypes  import c_byte, c_char, c_short, c_size_t, c_uint, c_ubyte, c_ulong, c_void_p, c_wchar
-from _ctypes import _SimpleCData
-from numpy   import array, empty, full
-import libpath
+from ctypes    import addressof, CDLL, memset, RTLD_GLOBAL, sizeof
+from ctypes    import c_bool, c_char_p, c_double, c_float, c_int, c_long, c_void_p, POINTER
+from ctypes    import c_byte, c_char, c_short, c_size_t, c_uint, c_ubyte, c_ulong, c_void_p, c_wchar
+from _ctypes   import _SimpleCData
+from numpy     import array, dtype, empty, full, ndindex
+from math      import ceil, prod
+from itertools import product
+from inspect   import stack
+from weakref   import WeakValueDictionary
+from time import time
+_accessor_cache = {}
+_DL_CACHE = WeakValueDictionary()
 selectcases = { float      : c_double,
                 c_double   : c_double,
                 c_float    : c_float,
@@ -51,10 +56,341 @@ typestrs = { c_double : 'f8',
              c_bool   : 'b4',
              c_char_p : 'S'
            }
+class _DL_DT_Scalar:
+    __slots__ = ('_parent', '_meta')
+    def __init__(self, parent, meta):
+        self._parent = parent
+        self._meta   = meta
+    def instantiate(self):
+        addr        = self._parent.__address__ + self._meta['_offset']
+        typename    = self._meta['_typename']
+        dict_dt     = self._parent._derived_types[typename]
+        caller_base = (self._parent._caller + '%' + self._meta['_name'] if self._parent._caller else self._meta['_name'])
+        inst = DL_DT(addr,
+                     dict_dt,
+                     self._parent._derived_types,
+                     caller=caller_base,
+                     caller_type=typename,
+                     return_ctype=self._parent._return_ctype,
+                     debug=self._parent._debug)
+        setattr(inst, '_derived_type', typename)
+        return inst
+class _DL_DT_Pointer:
+    __slots__ = ('_parent', '_meta')
+    def __init__(self, parent, meta):
+        self._parent = parent
+        self._meta   = meta
+    def instantiate(self):
+        offset   = self._meta['_offset']
+        typename = self._meta['_typename']
+        dict_dt  = self._parent._derived_types[typename]
+        addr = c_ulong.from_address(self._parent.__address__ + offset).value
+        if addr == 0:
+            return
+        if self._parent.lib_dl_py2f.associated(c_ulong(addr)):
+            return
+        caller      = self._parent._caller
+        name_member = self._meta['_name']
+        caller_base = (caller + '%' + name_member) if caller else name_member
+        is_linked_list = False
+        if caller:
+            base = caller.split('%')[-1].split('(')[0]
+            is_linked_list = (base == name_member)
+            if is_linked_list and self._parent._debug:
+                print(f' >>> NB: {_bE}{caller_base}{CLR_} is a {_bW}scalar linked list!{CLR_}')
+        if self._parent.lib_dl_py2f.associated(c_ulong(addr)):
+            if self._parent._debug:
+                print(f' >>> WARNING: {_R}{caller_base}{CLR_} at {_uW}{hex(addr)}{CLR_} is likely to be a wild pointer!')
+            return
[…]
