### 2025-12-06 17a1019
Revised paper.md in response to the JOSS reviewer's comments, after a major upgrade of the code
