+++ paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-Fortran is long established as one of the major programming languages for scientific software programs, but has only limited facilities for interoperating with other modern languages such as Python. `DL_PY2F` is an open-source library for the creation of modern interfaces and data structures in Python that can interoperate with existing scientific software written in Fortran and manipulate their data.
+Fortran is long established as one of the major programming languages for scientific software, but has only limited facilities for interoperating with other modern languages such as Python. `DL_PY2F` is an open-source library for the creation of modern interfaces and data structures in Python that can interoperate with existing scientific software written in Fortran and manipulate their data.
 
 # Statement of need
 
@@ -39,7 +39,7 @@ While the `DL_PY2F` library is vital for the `Py-ChemShell` program, it has the 
     libapp = ctypes.CDLL('/abc/def/libapp.so')
     ierror = libapp.interface_app(dl_py2f.py2f(appObj))
 
-Here, `appObj` is an instance, typically created by the user at runtime, of the new package's Python class `App` which inherits from `ctypes.Structure`, for example:
+Here, `appObj` is an instance, typically created by the application's user at runtime, of the new package's Python class `App` which inherits from `ctypes.Structure`, for example:
 
     import ctypes, numpy
     from . import callback
@@ -62,7 +62,7 @@ On the Fortran side, the `interface_app` function receives the passed-in `appPtr
                 use iso_c_binding
             endfunction callback
         endinterface
-        type(dictType)     , pointer, public :: PyApp
+        type(dictType)     , pointer, public :: PyApp, PyChild
         procedure(callback), pointer, public :: PyCallback
         contains
         function interface_app(appPtr) bind(c) result(ireturn)
@@ -71,39 +71,41 @@ On the Fortran side, the `interface_app` function receives the passed-in `appPtr
             type(c_funptr)                      :: pyfuncPtr
             type(PyType)  , pointer             :: childPtr
             real(kind=8)  , pointer             :: coords(:,:)
-            integer                             :: npoints
             allocate(PyApp, source=ptr2dict(appPtr))
             call PyApp%get('child', childPtr)
             allocate(PyChild, source=ptr2dict(childPtr))
-            allocate(coords(3,npoints))
-            call PyApp%get('coords', coords)
             call PyApp%get('callback', pyfuncPtr)
             call c_f_procpointer(pyfuncPtr, PyCallback)
-            ! run the application
-            call my_app(npoints, coords)
-            call PyApp%set('coords', coords)
-            deallocate(PyApp, PyChild, coords)
+            call PyApp%get('coords', coords) ! a handle with write access to the NumPy array
+            call my_app(coords)              ! run the application
+            deallocate(PyApp, PyChild)
+            nullify(coords)                  ! cannot be deallocated as it does not own the NumPy array
         endfunction interface_app
     endmodule AppModule
 
-A great advantage of `DL_PY2F` for the application developers is that the attributes of the Python instance are conveniently retrieved by querying their names in a dictionary-like way. Two type-bound procedures `get` and `set` grant read and write access, respectively. For arrays, no copies are made because both operations act through memory addresses and values are thus changed in place and reflected on the Python side (if mutable). Callback functions can also be invoked, as follows:
[…]
