### 2025-12-07 c9f763f
Made necessary modifications to release via PyPI
- Removed from colours import *
- Adopted importing from relative path
