+++ dl_f2py.py
@@ -19,8 +19,8 @@
 # you.lu@stfc.ac.uk
 __email__ = 'you.lu@stfc.ac.uk'
 __author__ = f'You Lu <{__email__}>'
-import libpath, os, utils
-from utils.colours import *
+import os, utils
+from importlib.resources import files
 from ctypes    import addressof, CDLL, memset, RTLD_GLOBAL, sizeof
 from ctypes    import c_bool, c_char_p, c_double, c_float, c_int, c_long, c_void_p, POINTER
 from ctypes    import c_byte, c_char, c_short, c_size_t, c_uint, c_ubyte, c_ulong, c_void_p, c_wchar
@@ -30,7 +30,9 @@ from math      import ceil, prod
 from itertools import product
 from inspect   import stack
 from weakref   import WeakValueDictionary
-from time import time
+from time      import time
+from .         import libpath
+from .utils    import colours as _clr
 _accessor_cache = {}
 _DL_CACHE = WeakValueDictionary()
 selectcases = { float      : c_double,
@@ -97,10 +99,10 @@ class _DL_DT_Pointer:
             base = caller.split('%')[-1].split('(')[0]
             is_linked_list = (base == name_member)
             if is_linked_list and self._parent._debug:
-                print(f' >>> NB: {_bE}{caller_base}{CLR_} is a {_bW}scalar linked list!{CLR_}')
+                print(f' >>> NB: {_clr._bE}{caller_base}{_clr.CLR_} is a {_clr._bW}scalar linked list!{_clr.CLR_}')
         if self._parent.lib_dl_py2f.associated(c_ulong(addr)):
             if self._parent._debug:
-                print(f' >>> WARNING: {_R}{caller_base}{CLR_} at {_uW}{hex(addr)}{CLR_} is likely to be a wild pointer!')
+                print(f' >>> WARNING: {_clr._R}{caller_base}{_clr.CLR_} at {_clr._uW}{hex(addr)}{_clr.CLR_} is likely to be a wild pointer!')
             return
         inst = DL_DT(addr,
                     dict_dt,
@@ -115,7 +117,7 @@ class _DL_DT_Pointer:
                 self._parent._return_immediately = True
             else:
                 if self._parent._debug:
-                    print(f' >>> WARNING: {_R}{caller_base}{CLR_} is likely to be an unassociated pointer to a linked list! (type: {_Y}{typename}{CLR_})')
+                    print(f' >>> WARNING: {_clr._R}{caller_base}{_clr.CLR_} is likely to be an unassociated pointer to a linked list! (type: {_clr._Y}{typename}{_clr.CLR_})')
             return inst
         return inst
 class _DL_DT_Array:
@@ -413,7 +415,8 @@ class DL_DT(_SimpleCData):
         try:
             return self._lib_dl_py2f
         except:
-            self._lib_dl_py2f = utils.modutils.getLinkedLib('dl_py2f', libpath.libpath, is_linked=True)
+            self._lib_dl_py2f = utils.modutils.getLinkedLib('dl_py2f', files(__package__), is_linked=True) \
+                             or utils.modutils.getLinkedLib('dl_py2f', libpath.libpath, is_linked=True)
             return self._lib_dl_py2f
     def __init__(self,
                  address      :int,
@@ -459,8 +462,8 @@ class DL_DT(_SimpleCData):
                         if lbuff[2] != v['_size_chunk']:
                             _is_allocated = False
                             if debug:
-                                print(f'{_bR}\n >>> WARNING: = In {CLR_}{_bY}{caller}{CLR_}{_bR}, there may be something wrong with derived-type array {_bY}{k}{CLR_}{_bR}!{CLR_}')
-                                print(f'{_bR}              Shape: {shape}{CLR_}')
[…]
