### 2025-12-08 e1476b2
Updated README.md and paper.md to reflect the release on PyPI
