+++ README.md
@@ -76,7 +76,7 @@ Alternatively, `DL_PY2F` can be installed via PyPI (the `pip` command):
 
 `pip install dl_py2f`
 
-:caution: **Caution:** `DL_PY2F` deployed via PPA (`apt`) and PyPI (`pip`) only works with application projects compiled using gfortran due to dl_py2f.mod we shipped.
+:heavy_exclamation_mark: **Caution:** `DL_PY2F` deployed via PPA (`apt`) and PyPI (`pip`) only works with application projects compiled using gfortran due to dl_py2f.mod we shipped.
 
 The use of system installed `DL_PY2F` is also demonstrated in the [example](https://github.com/stfc/dl_py2f-example) package.
 
