### 2025-12-08 b427a07
Prefixed the Linux commands with $
