### 2026-01-08 c0666b7
Slightly modified README.md and paper.md to reflect the latest update of PyPI deployment which allows the user (app developer) to compile DL_PY2F in place when installing via pip
