### 2026-01-16 stfc/dl_py2f#35 @mgoonde
[JOSS REVIEW #2]
Hello,
this is the second round of review at https://github.com/openjournals/joss-reviews/issues/8992
apologies for my delay.

All concerns raised in the first review (#19) have been properly addressed.

The concern regarding data copies now seems fine, although the wording would probably be more direct/transparent to say `copy` instead of `keepref`/`readonly`, but this is really just personal choice, and should not have an impact on the work.

I have tested the reviewed code on two machines. On my laptop the example previously took 12s, after revision it takes 6s, and produces some error during the run:
```python
 >>> callback (Py): scanning along var_of_t04.t2ar02[:,2], ivar04 =
1004 1004 1004 1004 1004 1004
 >>> callback (Py): scanning along var_of_t04.t2ar02[1:5,:], ivar04 =
Exception ignored on calling ctypes callback function: <function callback at 0x747f7dc3a200>
Traceback (most recent call last):
  File "/home/mgunde/dl_py2f-example_resub/example/callback.py", line 242, in callback
    print(a.ivar04, end=' ')
AttributeError: 'numpy.ndarray' object has no attribute 'ivar04'
```

On a different machine however I do see the claimed speedup, without any errors in the provided example code. Thus most probably the issue is from my laptop (probably related to version of some dependency, maybe numpy?).

Some bullet points from the JOSS review checklist do not seem to be satisfied though:

- There does not seem to be a mechanism for checking dependency installation/verification;
- The only way one can understand how to use the code is by working through the example page and the actual code, no other documentation seems to be available;
- The code itself is practically not commented, no help strings (docstrings) are available for the public routines/functions;
