+++ dl_f2py.py
@@ -22,7 +22,7 @@ __author__ = f'You Lu <{__email__}>'
 import os, utils
 from importlib.resources import files
 from ctypes    import addressof, CDLL, memset, RTLD_GLOBAL, sizeof
-from ctypes    import c_bool, c_char_p, c_double, c_float, c_int, c_long, c_void_p, POINTER
+from ctypes    import c_bool, c_char_p, c_double, c_float, c_int, c_long, c_void_p, POINTER, Structure
 from ctypes    import c_byte, c_char, c_short, c_size_t, c_uint, c_ubyte, c_ulong, c_void_p, c_wchar
 from _ctypes   import _SimpleCData
 from numpy     import array, dtype, empty, full, ndindex
@@ -35,6 +35,10 @@ from .         import libpath
 from .utils    import colours as _clr
 _accessor_cache = {}
 _DL_CACHE = WeakValueDictionary()
+class c_complex8(Structure):
+    _fields_ = [('re', c_double), ('im', c_double)]
+class c_complex4(Structure):
+    _fields_ = [('re', c_float), ('im', c_float)]
 selectcases = { float      : c_double,
                 c_double   : c_double,
                 c_float    : c_float,
@@ -43,6 +47,8 @@ selectcases = { float      : c_double,
                 c_long     : c_long,
                 str        : c_char_p,
                'CHARACTER1': c_char_p,
+               'COMPLEX4'  : c_complex4,
+               'COMPLEX8'  : c_complex8,
                'INTEGER4'  : c_int,
                'INTEGER8'  : c_long,
                'REAL4'     : c_float,
@@ -51,12 +57,14 @@ selectcases = { float      : c_double,
                'LOGICAL4'  : c_bool, # TODO logical of other lengths
                'LOGICAL8'  : c_bool, # TODO logical of other lengths
               }
-typestrs = { c_double : 'f8',
-             c_float  : 'f4',
-             c_long   : 'i8',
-             c_int    : 'i4',
-             c_bool   : 'b4',
-             c_char_p : 'S'
+typestrs = { c_double  : 'f8',
+             c_float   : 'f4',
+             c_long    : 'i8',
+             c_int     : 'i4',
+             c_bool    : 'b4',
+             c_char_p  : 'S',
+             c_complex4: 'c8',
+             c_complex8: 'c16',
            }
 class _DL_DT_Scalar:
     __slots__ = ('_parent', '_meta')
@@ -441,8 +449,6 @@ class DL_DT(_SimpleCData):
             if not isinstance(v, dict) or '_offset' not in v:
                 continue
             icount += 1
-#            if type(v) is dict:
-#                if '_offset' in v:
             if '_ndims' in v:
                 if v.get('_is_derived', False):
                     if v.get('_is_deferred', False):
@@ -507,32 +513,6 @@ class DL_DT(_SimpleCData):
                                               '_ndims'     :  v.get('_ndims', 0),
                                               '_is_pointer':  v.get('_is_pointer', False)}
                         continue
-#                        value = DL_DT(self.__address__+offset,
-#                                      derived_types[v['_type']],
-#                                      derived_types,
-#                                      caller=caller+'%'+k,
-#                                      caller_type=v['_type'],
-#                                      return_ctype=return_ctype,
-#                                      debug=debug)
-#                        setattr(value, '_derived_type', v['_type'])
-#                        value.__array__ = empty(shape=shape[::-1], dtype=object)
-#                        setattr(self, k, value)
-#                        indices = product(*dims)
-#                        if debug:
-#                            for i, ind in enumerate(indices):
-#                                lcallers += [ caller+'%'+k+'('+','.join([str(_+1) for _ in ind[::-1]])+')' ]
-#                        else:
-#                            lcallers = [ caller ]*len(list(product(*dims)))
[…]
