### 2026-01-18 211ec4c
Fixed regression test broken by recently added stuff
