+++ dl_f2py.py
@@ -179,7 +179,7 @@ class _DL_DT_Array:
         if self._meta['_kind'] == 'array_deferred':
             base = c_ulong.from_address(self._parent.__address__ + offset).value
             if base == 0:
-                raise RuntimeError(f'Deferred array "{self._meta["_name"]}" is not allocated (NULL pointer).')
+                raise RuntimeError(f' >>> DL_F2PY ERROR: Deferred array "{self._meta["_name"]}" is not allocated (NULL pointer).')
         else:
             base = self._parent.__address__ + offset
         addr = base + flat*size_chunk
@@ -435,8 +435,7 @@ class DL_DT(_SimpleCData):
                  return_ctype :bool = False,
                  debug        :bool = False):
         if type(dict_dt) is not dict or not dict_dt.get('_is_derived', False):
-            print('\n >>> DL_F2PY ERROR: Argument dict_dt must be a dict containing information of data structure of a Fortran derived type', flush=True)
-            exit(999)
+            raise RuntimeError(' >>> DL_F2PY ERROR: Argument dict_dt must be a dict containing information of data structure of a Fortran derived type', flush=True)
         self._dict_dt       = dict_dt
         self._derived_types = derived_types
         self._caller        = caller
@@ -633,6 +632,7 @@ class DL_DT(_SimpleCData):
                             sbuff_setter = ''
                             if v['_type'] == '_tb_procedure':
                                 continue
+                            addr = self.__address__ + v['_offset']
                             value = v['_type'].from_address(self.__address__+v['_offset'])
                             setter_args = f'self._DL_DT_{k}'
                             setattr(self, '_DL_DT_'+k, value)
@@ -894,11 +894,14 @@ class DL_DL(CDLL):
                 for moddir in self.moddir:
                     self.parseAllModules(moddir, caller='modules')
             return self._modules
-        except:
-            self._modules = self._DL_MOD(self, {'_name': '__modules__'})
-            if self._to_parse_all:
-                for moddir in self.moddir:
-                    self.parseAllModules(moddir, caller='modules')
+        except Exception:
+            import traceback, sys
+            traceback.print_exc()
+            if not hasattr(self, '_modules'):
+                self._modules = self._DL_MOD(self, {'_name': '__modules__'})
+            self._to_parse_all = False
+            for moddir in self.moddir:
+                self.parseAllModules(moddir, caller='modules')
             return self._modules
     @modules.setter
     def modules(self, val):
@@ -908,7 +911,8 @@ class DL_DL(CDLL):
                 print(f'\n >>> WARNING: The following module instances will be overwritten:\n             ', ', '.join(overlap))
             self._modules.update(val)
         except:
-            self._modules = self._DL_MOD(self, {'_name': '__modules__'})
+            if not hasattr(self, '_modules'):
+                self._modules = self._DL_MOD(self, {'_name': '__modules__'})
             self._modules.update(val)
     @property
     def nsymbols(self):
@@ -921,14 +925,13 @@ class DL_DL(CDLL):
             symbols_module = self.getModuleSymbols(module)
             lbuff = [ s for s in symbols_module if s.lower().rstrip('_').endswith('_'+symbol.strip().lower()) ]
             if len(lbuff) == 0:
-                print(f' >>> DL_PY2F ERROR: Module {module} does not have symbol "{symbol}"\n', flush=True)
-                exit(999)
+                raise RuntimeError(f' >>> DL_PY2F ERROR: Module {module} does not have symbol "{symbol}"\n', flush=True)
             elif len(lbuff) == 1:
                 return lbuff[0]
             else:
-                print(f' >>> DL_PY2F ERROR: Module {module} has more than one symbols containing string "{symbol}":')
[…]
