### 2026-01-21 5677e73
Added support for Flang/Clang++ and updated README.md and paper.md accordingly
