+++ __init__.py
@@ -21,7 +21,7 @@ __author__     = [ 'You Lu', 'Thomas W. Keal' ]
 __copyright__  = 'Copyright (C) 2025 The authors of DL_PY2F'
 __credits__    = [ 'You Lu', 'Thomas W. Keal' ]
 __license__    = 'LGPLv3'
-__version__    = '1.1.3'
+__version__    = '1.1.4'
 __maintainer__ = __author__[0]
 __email__      = 'you.lu@stfc.ac.uk'
 __status__     = ''
