### 2026-02-02 b11639e
Yeah! Made gvfortran essentially work!
