+++ dl_py2f.f90
@@ -17,6 +17,7 @@
 !  <https://www.gnu.org/licenses/>.
 
 ! you.lu@stfc.ac.uk
+
 module DL_PY2F
     use iso_c_binding
     implicit none
@@ -42,9 +43,11 @@ module DL_PY2F
         type(CStructType) :: cdata
     endtype PyType
     type dictType
-        character(len=LENSTR)     :: type  = ""
-        integer                   :: sizem =  0
-        integer                   :: sizen =  0
+        character(len=LENSTR)     :: type   = ""
+        integer                   :: sizem  =  0
+        integer                   :: sizen  =  0
+        integer                   :: stride =  1
+        type(c_ptr)               :: addr0
         character(len=:), pointer :: key             => null()
         class(*)        , pointer :: scalar          => null()
         class(*)        , pointer :: array(:)        => null()
@@ -58,6 +61,7 @@ module DL_PY2F
         real(kind=8)    , pointer :: twodimdbl(:,:)  => null()
         real(kind=8)    , pointer :: onedimcdbl(:)   => null()
         real(kind=8)    , pointer :: twodimcdbl(:,:) => null()
+        type(c_funptr)  , pointer :: cfuncptr        => null()
         type(dictType)  , pointer :: next            => null()
         contains
         generic, public :: assign => assignScalar,     &
@@ -104,6 +108,7 @@ module DL_PY2F
         procedure, private :: assignOneDimLng
         procedure, private :: assignTwoDimLng
         procedure, private :: assignPyPtr
+        procedure, private :: assignCFuncPtr
         procedure, private :: getIntCLong
         procedure, private :: getInt
         procedure, private :: getReal
@@ -191,10 +196,10 @@ module DL_PY2F
                     call metaobj%assign(trim(namebuff), nonetype)
                 case('function')
                     call c_f_pointer(cdata(i)%attr, cfuncbuff)
-                    call metaobj%assign(trim(namebuff), cfuncbuff)
+                    call metaobj%assignCFuncPtr(trim(namebuff), cfuncbuff)
                 case('object')
                     call c_f_pointer(cdata(i)%attr, pyptrbuff)
-                    call metaObj%assign(trim(namebuff), pyptrbuff)
+                    call metaObj%assignPyPtr(trim(namebuff), pyptrbuff)
                 case('int')
                     call c_f_pointer(cdata(i)%attr, cintbuff)
                     call metaObj%assign(trim(namebuff), cintbuff)
@@ -248,7 +253,8 @@ module DL_PY2F
                                 endif
                                 if(sizen.eq.1) then
                                     onedimdblbuff => dblarraybuff(1,1:sizem)
-                                    call metaObj%assign(trim(namebuff), onedimdblbuff, keepReference)
+                                    call metaObj%assign(trim(namebuff), onedimdblbuff, keepReference, &
+                                                        cdata(i)%attr, ncols)
                                 elseif(sizen.gt.1) then
                                     twodimdblbuff => dblarraybuff(1:sizen,:)
                                     call metaObj%assign(trim(namebuff), twodimdblbuff, keepReference)
@@ -257,7 +263,8 @@ module DL_PY2F
                                 call c_f_pointer(cdata(i)%attr, dblarraybuff, (/1, sizem*sizen/))
                                 if(sizen.eq.1) then
                                     onedimdblbuff => dblarraybuff(1,:)
-                                    call metaObj%assign(trim(namebuff), onedimdblbuff, keepReference)
+                                    call metaObj%assign(trim(namebuff), onedimdblbuff, keepReference, &
+                                                        cdata(i)%attr, 1)
                                 elseif(sizen.gt.1) then
[…]
