+++ README.md
@@ -23,7 +23,7 @@ You Lu and Thomas W. Keal, *Journal of Open Source Software*, in preparation
 
 ## Project status
 
-The Python-to-Fortran interoperability has been comprehensively tested using GNU, Intel, and Flang/Clang++ compilers.
+The Python-to-Fortran interoperability has been comprehensively tested using GNU, Intel, and Flang/Clang++ compilers. It also works with NVIDIA HPC compilers although currently integer arrays must be retrieved and altered in the "safe mode" (`readonly=.true.`) due to a bug in nvfortran.
 
 :warning: **Warning:** However, the method for Fortran-to-Python interoperability is still undergoing testing and validation, and is currently limited to use with the GNU compiler **gfortran**, as the proprietary .mod file format used by the Intel compiler or other compilers' .mod format is not yet supported.
 
@@ -38,6 +38,7 @@ The Python-to-Fortran interoperability has been comprehensively tested using GNU
 | OR icpc/ifort                 | 17           | :warning: |
 | OR icpx/ifx                   | 2024         | :warning: |
 | OR clang++/flang              | 22           | :warning: |
+| OR nvc++/nvfortran            | 26.1         | :warning: |
 | cmake                         | 3.16         |      |
 | python3-dev                   | 3.8          |      |
 | python3-numpy                 | 1.21.5       |      |
@@ -45,8 +46,9 @@ The Python-to-Fortran interoperability has been comprehensively tested using GNU
 :bulb: The above package names are based on Ubuntu Linux. They may vary on other
        operating systems.
 
-:warning: The Fortran-to-Python method does **NOT** yet work with the Intel compilers as Intel's proprietary
-          .mod file format is unpublished and unsupported. Its support for Flang/Clang++ is yet to be implemented.
+:warning: The Fortran-to-Python method does **NOT** yet work with the Intel, Flang/Clang++, or NVIDIA compilers.[^a]
+
+[a]: Intel's proprietary .mod file format is unpublished and unsupported. Support for Flang/Clang++ and NVIDIA compilers is yet to be implemented.
 
 ## Using and testing `DL_PY2F`
 
@@ -80,7 +82,11 @@ or Flang/Clang++
 
 `export FC=flang-22; export CXX=clang++-22`
 
-:warning: Please keep in mind that only the Python-to-Fortran interoperability works with Intel compilers.
+or NVIDIA HPC compilers
+
+`export FC=nvfortran; export CXX=nvc++`
+
+:warning: Please keep in mind that only the Python-to-Fortran interoperability works with non-GNU compilers.
 
 `$ python3 -m venv .venv`
 
