### 2026-02-02 2700e5a
Update paper.md
