### 2026-02-02 7600c92
Update paper.md
