### 2026-02-03 add04ec
Corrected compilation output dir; updated version number
