### 2026-02-03 22f2fcc
Removed libpath.py.in
