+++ libpath.py.in
@@ -1,1 +1,0 @@
-libpath = '@CMAKE_LIBRARY_OUTPUT_DIRECTORY@'
