### 2026-02-22 6a03844
Adopted recent improvements in __init__.py
- Cached subclasses created from CStructure to increase performance and reduce memory use
- Bypassed unsupported Python objects
- Inserted %finalise to the code snippet in paper.md
