### 2026-03-21 #8992 @chemsh
> hi [@danielskatz](https://github.com/danielskatz), thanks for checking in following up with the reviewers.
>
> [@awvwgk](https://github.com/awvwgk) just checking in on the review here.
>
> [@mgoonde](https://github.com/mgoonde) just checking in on the review here.
>
> [@chemsh](https://github.com/chemsh) feel free to start addressing [@mgoonde](https://github.com/mgoonde)'s comments.

  Dear @HaoZeke, we would like to inform you that we have addressed the documentation concerns raised by the reviewer. A comprehensive [REFERENCE_MANUAL](https://github.com/stfc/dl_py2f/blob/main/REFERENCE_MANUAL.md) has been added to the repository, covering installation, supported data types, memory ownership, MPI patterns, the full API, and known limitations. It's also mentioned in paper.md. A detailed response has also been posted in the reviewer's issue thread.
