### 2026-04-29 #8992 @chemsh
Hi @HaoZeke and @danielskatz, brief update from our side:
Since the v1.0 submission, three further releases have shipped on PyPI: v1.2.0 (nvfortran F2P), v1.3.0 (Flang/LLVM F2P), and v1.4.0 (Intel ifx F2P). A REFERENCE_MANUAL was also added in v1.2.0. Reviewer @mgoonde signed off positively on 2026-03-03 (#35). We remain available to address any remaining concerns from @awvwgk or to discuss next steps if a replacement reviewer would help move things forward. Thanks for your patience.
